// resources/lib/generated/api/types.gen.ts
export type DashboardSummary = {
    createdAt: string;
    id: string;
    isDefault: boolean;
    isOwner: boolean;
    isShared: boolean;
    itemCount?: number | null;
    name: string;
    ownerName?: string | null;
    position: number;
    visibility: DashboardVisibility;
};
// ...
export type DashboardVisibility = 'private' | 'team';